edit

+++ printf/ft_print_pointer.c
@@ -6,7 +6,7 @@
 /*   By: hkubo <hkubo@student.42tokyo.jp>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/03/18 18:41:01 by hkubo             #+#    #+#             */
-/*   Updated: 2021/03/21 10:07:13 by hkubo            ###   ########.fr       */
+/*   Updated: 2021/03/21 10:24:39 by hkubo            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -67,7 +67,7 @@ int print_pointer(va_list *ap, a_list flag_list)
     str_num = ft_itoa_hex_long(num, "0123456789abcdef");
     // if (num == 0 && flag_list.precision == 0)
     //     str_num = "";
-    keta = ft_strlen(str_num);
+    keta = (int)ft_strlen(str_num);
     // printf("keta: %d\n", keta);
     size = 0;
     if (flag_list.precision == 0 && num == 0)
